Added release, Buff functions

+++ Revive/Revive/View Model/ReviveManager/ReviveManager+Items.swift
@@ -8,17 +8,6 @@
 import Foundation
 
 extension ReviveManager {
-    func updateItemAmount(id: Int, newAmount: Int) {
-        DataManager.shared.updateItemCurrTimeData(for: id, with: newAmount, myItems: myItems)
-    }
-    
-    func saveNewItem(id: Int, amount: Int) {
-        DataManager.shared.saveItemData(customItem: MyItems(id: id, amount: amount))
-    }
-    
-    func deleteItem(id: Int) {
-        DataManager.shared.deleteItemData(for: id)
-    }
     
     func getItem(id: Int) -> Item {
         return itemList[id - 2001]
@@ -43,16 +32,20 @@ extension ReviveManager {
         for i in 0..<getItem(id: currPanelItem!.id).functionType.count {
             switch getItem(id: currPanelItem!.id).functionType[i].rawValue {
             case "hatching_time":
-                continue
+                ItemsforHatchingTime(index: i)
+                useItem(save: false)
             case "training_time":
-                continue
+                ItemsforTrainingTime(index: i)
+                useItem(save: false)
             case "exploring_time":
-                continue
+                ItemsforExploringTime(index: i)
+                useItem(save: false)
             case "rarity_up":
-                continue
+                RarityUp(index: i)
+                useItem(save: false)
             case "exp":
                 ItemsAddEXP(index: i)
-                useItem()
+                useItem(save: true)
                 
             default:
                 continue
@@ -60,21 +53,48 @@ extension ReviveManager {
         }
     }
     
-    func useItem() {
+    func updatePendingItem() {
+        for item in pendingItem {
+            if item.value == 0 {
+                deleteItem(id: item.key)
+                myItems.remove(at: getMyItemIndex(id: item.key))
+            } else {
+                updateItemAmount(id: item.key, newAmount: myItems[getMyItemIndex(id: item.value)].amount)
+            }
+        }
+    }
+    
+    func resetBuffRate() {
+        isHatchingBuff = false
+        isTrainingBuff = false
+        isExploringBuff = false
+        isRarityBuff = false
+        rarityBuffRate = 0
+        hatchingBuffRate = 1
+        trainingBuffRate = 1
+        exploringBuffRate = 1
+    }
+    
+    func useItem(save: Bool) {
         myItems[getMyItemIndex(id: currPanelItem!.id)].amount -= 1
         currPanelItem = myItems[getMyItemIndex(id: currPanelItem!.id)]
         
-        updateItemAmount(id: currPanelItem!.id, newAmount: myItems[getMyItemIndex(id: currPanelItem!.id)].amount)
-        
-        if myItems[getMyItemIndex(id: currPanelItem!.id)].amount == 0 {
-            deleteItem(id: currPanelItem!.id)
-            myItems.remove(at: getMyItemIndex(id: currPanelItem!.id))
-            if myItems.count == 0 {
-                currPanelItem = nil
-            } else {
-                currPanelItem = myItems[0]
-            }
+        if save {
+            updateItemAmount(id: currPanelItem!.id, newAmount: myItems[getMyItemIndex(id: currPanelItem!.id)].amount)
+            
+            if myItems[getMyItemIndex(id: currPanelItem!.id)].amount == 0 {
+                deleteItem(id: currPanelItem!.id)
+                myItems.remove(at: getMyItemIndex(id: currPanelItem!.id))
+                if myItems.count == 0 {
+                    currPanelItem = nil
+                } else {
+                    currPanelItem = myItems[0]
+                }
 
+            }
+            
+        } else {
+            pendingItem[currPanelItem!.id] = currPanelItem!.amount
         }
     }
     
@@ -85,5 +105,25 @@ extension ReviveManager {
         mySpecies[getSpeciesIndex(id: currPanelSpecies?.speciesID ?? 0, date: currPanelSpecies?.hatchDate ?? Date())].currExp = currExp
         currPanelSpecies?.level += levelUpNum
         currPanelSpecies?.currExp = currExp
+    }
+    
+    func ItemsforHatchingTime(index: Int) {
+        isHatchingBuff = true
+        hatchingBuffRate *= Double(getItem(id: currPanelItem!.id).amount[index])
+    }
+    
+    func ItemsforTrainingTime(index: Int) {
+        isTrainingBuff = true
+        trainingBuffRate *= Double(getItem(id: currPanelItem!.id).amount[index])
+    }
+    
+    func ItemsforExploringTime(index: Int) {
+        isExploringBuff = true
+        exploringBuffRate *= Double(getItem(id: currPanelItem!.id).amount[index])
+    }
+    
+    func RarityUp(index: Int) {
+        isRarityBuff = true
+        rarityBuffRate += Double(getItem(id: currPanelItem!.id).amount[index])
     }
 }
